arm64/dts: shusky: Add device tree entry for PERST delay
After RC has initialized PCIe PHY and the reference clock is stable, it
asserts PERST to let the EP know that it can start to bring up its PHY
and start link training. The RC needs to wait for the EP to finish this
initialization. The perst-delay-us in usecs defines this delay.

Bug: 242463095
Signed-off-by: Sajid Dalvi <sdalvi@google.com>
Change-Id: I1b985d99164f97f34841bcd19434676f38f30938