kleaf: Add support for building shusky with kleaf
This creates the necessary bazel build files to compile the shusky build
target with kleaf.

One unique difference is we are using labels to select the kernel build
and SOC module for the build target. These labels are defined in
//private/google-modules/soc/gs as:

  * gs_kernel_build
  * gs_soc_module

and allow one to define the kernel_build and SoC module dependencies
within external modules to a generic label. For shusky builds, use
--config=shusky to set these labels to the zuma_shusky and
zuma_soc.shusky starlarks accordingly. This is defined in the
device.bazelrc file.

Change-Id: I63dea54eb3626f7f6e12b414c5775c415d1f738b
Signed-off-by: Will McVicker <willmcvicker@google.com>