kleaf: add KMI symbol list support
As part of ABI monitoring, we need to maintain a symbol list of all the
core kernel symbols the kernel modules use. This list is maintained in
the AOSP kernel tree. To update this, you can run this command:

  bazel run --config=shusky --config=fast \
    //private/devices/google/shusky:zuma_shusky_abi_update_symbol_list

When ABI monitoring is enabled in android14-5.15, after updating the
symbol list the ABI xml representation will also need to be updated.
This can be done with the command:

  bazel run --config=local //aosp:kernel_aarch64_abi_update

Change-Id: Ibb93faa62f07072b87ea2929de91407941879ddd
Signed-off-by: Will McVicker <willmcvicker@google.com>