arm64/configs: shusky: enable pKVM support for GSA
Even though S2MPU is the always-on domain of GSA block, it is affected
by the sleep sequence that sets the clock mux that feeds into
GSACTRL/GSACORE blocks to DIV_16 during sleep. This means that GSA S2MPU
is running at AON_CLOCK/16 which makes latency too high.

As a workaround, wake up the GSA processor during S2MPU work so that
the clocks are at the AON_CLOCK frequency. GSA will then sleep when the
wake lock is released.

Bug: 265407176
Change-Id: I641ae571e3981743596c6804813485a2b8d67ae9
Signed-off-by: Kenny Root <kroot@google.com>